successful tests on Debug/Release builds with Ninja on Linux/PopOS machine

+++ game/libraries/abilities/stealth/Stealth.hpp
@@ -2,7 +2,7 @@
 
 //* Forward declarations
 class IGraphicle;
-using size_t = unsigned long long;
+using size_t = long unsigned int;
 
 /*
 @ Construct shallow pointer to Stealth object
